5 de novembro

+++ Droid/Resources/activitys/ItensActivity.cs
@@ -36,9 +36,9 @@ namespace MeusPedidos.Droid
 				if (nome.Text.Trim() == "" && preco.Text.Trim() == "" && quantidade.Text.Trim() == ""){
 					Toast.MakeText(this, "Preencha os campos acima", ToastLength.Long).Show();
 				} else {
-					itensPedido.Add(new ItemPedido(nome.Text, float.Parse(preco.Text), float.Parse(quantidade.Text), new Pedido("eu")));
+					itensPedido.Add(new ItemPedido(nome.Text, float.Parse(preco.Text), int.Parse(quantidade.Text), new Pedido("eu")));
 					itens.Adapter = new AdapterItensMain (this, itensPedido);
-					total =  total + (float.Parse(preco.Text) * float.Parse(quantidade.Text));
+					total =  total + (float.Parse(preco.Text) * int.Parse(quantidade.Text));
 				}
 			};
